% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_updateRProfile.R
\name{appnameToPath}
\alias{appnameToPath}
\title{Convert an appname to a path on the remote}
\usage{
appnameToPath(appname)
}
\arguments{
\item{appname}{a valid shiny app name, or "~" to return home directory}
}
\value{
A path wrt ~
}
\description{
Convert an appname to a path on the remote
}
