% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPAT.R
\name{check_pat_expiry}
\alias{check_pat_expiry}
\title{Check if the Github PAT is going to expire soon}
\usage{
check_pat_expiry(pat = Sys.getenv("GITHUB_PAT"), minweeks = 2)
}
\arguments{
\item{pat}{The Personal Access Token}

\item{minweeks}{Warn if fewer than minweeks weeks left until expiry}
}
\description{
Displays a warning if so
}
\details{
Gets the PAT from GITHUB_PAT by default
}
