% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_setupserver.R
\name{create_remote_dir}
\alias{create_remote_dir}
\title{Create a directory on the remote if it doesn't already exist}
\usage{
create_remote_dir(session, dirname)
}
\arguments{
\item{session}{The session to use}

\item{dirname}{The directory name, relative to ~}
}
\description{
Create a directory on the remote if it doesn't already exist
}
