% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_setupserver.R
\name{does_directory_exist}
\alias{does_directory_exist}
\title{Check whether a directory exists in user's homedir on the remote machine}
\usage{
does_directory_exist(session, dirname)
}
\arguments{
\item{session}{The session to use}

\item{dirname}{The directory name, relative to ~}
}
\description{
Check whether a directory exists in user's homedir on the remote machine
}
