% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_uploadappdir.R
\name{prepareRprofile}
\alias{prepareRprofile}
\title{Prepare a .Rprofile for staging}
\usage{
prepareRprofile(our_profile)
}
\arguments{
\item{our_profile}{Character vector containing the .Rprofile you wish to
modify}
}
\value{
A character vector containing the (potentially) modified .Rprofile
}
\description{
This function takes an .Rprofile and
modifies it with any proxy setup that is required, using the
local environment variables
}
