% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processls.R
\name{process_raw}
\alias{process_raw}
\title{Convert the output from an command executed on the remote
server into a vector; one row per line}
\usage{
process_raw(inraw)
}
\arguments{
\item{inraw}{A raw string of bytes (typically as output by
ssh::ssh_exec_internal()$stdout)}
}
\value{
A vector containing the returned data
}
\description{
Convert the output from an command executed on the remote
server into a vector; one row per line
}
