% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processls.R
\name{processls}
\alias{processls}
\title{Convert the output from an ls command exectuted on the remote
server into a vector of directory entries}
\usage{
processls(inraw)
}
\arguments{
\item{inraw}{A raw string of bytes (typically as output by
ssh::ssh_exec_internal()$stdout)}
}
\value{
A list of directory entries
}
\description{
Convert the output from an ls command exectuted on the remote
server into a vector of directory entries
}
