% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_updateRProfile.R
\name{send_file}
\alias{send_file}
\title{Send a file to the remote server}
\usage{
send_file(
  session,
  file,
  appname = NULL,
  remotepath = "~",
  name = basename(file)
)
}
\arguments{
\item{session}{The session to use}

\item{file}{local file path}

\item{appname}{The app to send the file to in `~/ShinyApps/`}

\item{remotepath}{A location to send the file to wrt ~. Exactly one of `appname` or `remotepath` must be specified.}

\item{name}{remote file name, defaults to `basename(file)`}
}
\description{
Send a file to the remote server
}
