% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_updateRProfile.R
\name{send_Rprofile}
\alias{send_Rprofile}
\title{Send an Rprofile file to the remote server}
\usage{
send_Rprofile(session, appname = "~", remotepath = NULL, Rprofile)
}
\arguments{
\item{session}{The session to use}

\item{appname}{The app to send  to send the  .Rprofile to in ~/ShinyApps/}

\item{remotepath}{A location to send the file to wrt ~.  Exactly one of appname or remotepath must be specified}

\item{Rprofile}{A vector containing the Rprofile to send}
}
\description{
Send an Rprofile file to the remote server
}
