% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_updateRProfile.R
\name{shinysenderize_Rprofile}
\alias{shinysenderize_Rprofile}
\title{Update a `.Rprofile` file with the `ShinySender` code needed to use
`renv` with the hosted applications}
\usage{
shinysenderize_Rprofile(
  Rprofile,
  action = c("update", "delete"),
  rprofilefragmentpath = ShinySenderRprofilePath()
)
}
\arguments{
\item{Rprofile}{A character vector containing the Rprofile}

\item{action}{Whether to update the remote Rprofile with the `shinysender`
fragment, or delete it}

\item{rprofilefragmentpath}{The path to the fragment to add}
}
\value{
The modified Rprofile, as a character vector
}
\description{
Based on the code used in `packrat:::editRprofileAutoloader()`
}
\details{
`Rprofile` should be the contents of the `.Rprofile` file, as a characater
vector (one entry per line). This is what you get from `ReadLines()` if using
a local `Rprofile`, or `shinysender:::process_raw()` if using the results of a
remote ssh command
}
