% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_updateRProfile.R
\name{ShinySenderRprofilePath}
\alias{ShinySenderRprofilePath}
\title{Returns the location of our .Rprofile fragment to install on remote
server}
\usage{
ShinySenderRprofilePath()
}
\value{
The location of our .Rprofile fragment
}
\description{
Returns the location of our .Rprofile fragment to install on remote
server
}
