% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_updateRProfile.R
\name{ShinySenderRprofilePathStaging}
\alias{ShinySenderRprofilePathStaging}
\title{Returns the location of our staging .Rprofile fragment to install on remote
server}
\usage{
ShinySenderRprofilePathStaging()
}
\value{
The location of our .Rprofile fragment
}
\description{
Returns the location of our staging .Rprofile fragment to install on remote
server
}
