% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_listapps.R
\name{ss_appreport}
\alias{ss_appreport}
\title{Report on the apps installed on the remote server}
\usage{
ss_appreport(session, appdir = "~/ShinyApps")
}
\arguments{
\item{session}{The ssh session}

\item{appdir}{The location of the folder containing the user's ShinyApps}
}
\value{
A data frame containing details about each entry in the user's app
directory
}
\description{
Return a tibble containing details of each directory in the user's app
directory:
}
\details{
* Is the directory a potentially valid Shiny app (i.e. contains app.R or (ui.R and server.R))
* Does it have a packrat directory?
* Any logs associated with the app
}
