% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_listapps.R
\name{ss_appreport_linked}
\alias{ss_appreport_linked}
\title{Hyperlinked app report}
\usage{
ss_appreport_linked(session, slug, appdir = "~/ShinyApps")
}
\arguments{
\item{session}{The ssh session}

\item{slug}{The location of the hosted app, up to the appname}

\item{appdir}{The location of the folder containing the user's ShinyApps}
}
\value{
A data frame containing details about each entry in the user's app
directory with clickable links
}
\description{
Provide an app report, with html entries for entrynames, linking to the hosted app
}
