% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_bundleapp.R
\name{ss_bundleapp}
\alias{ss_bundleapp}
\title{Prepare an application bundle}
\usage{
ss_bundleapp(appDir = ".", appName)
}
\arguments{
\item{appDir}{The directory containing the Shiny app. Defaults to current
directory}

\item{appName}{The name to call the app when it is published}
}
\value{
The location of the created bundle file.
}
\description{
Prepare an application bundle, containing the app code, data and
package dependencies in packrat form
}
\details{
This function is a wrapper around some of the (internal) functionality
of [rsconnect][pkg::rsconnect]
}
