% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_connect.R
\name{ss_connect}
\alias{ss_connect}
\title{Connect to a Shiny Server}
\usage{
ss_connect(username, keyfile = NULL, config = "default", hostverify = TRUE)
}
\arguments{
\item{username}{The user to log onto the server as}

\item{keyfile}{The path to the user's private keyfile.}

\item{config}{The configuration name to use from config.yml. This will
follow the config package's inheritance rules (i.e. default: must exist
in the file, other configs inherit values not explicitly specified)}

\item{hostverify}{If TRUE, abort if the host's expected SHA1 isn't given in
the active config.  Will warn otherwise.}
}
\value{
An ssh connection object
}
\description{
Connect to the Shiny server referred to in the specified
configuration in config.yml, using the given user name
and keyfile specified
}
\note{
If keyfile is NULL, we use ssh::ssh_connects() approach to authentication
i.e. .ssh/id_rsa, followed by interactive password authentication
}
