% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_setupserver.R
\name{ss_create_shinyapps}
\alias{ss_create_shinyapps}
\title{Create ~/ShinyApps on remote account}
\usage{
ss_create_shinyapps(session)
}
\arguments{
\item{session}{The session to use}
}
\description{
Create ~/ShinyApps on remote account
}
