% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_deleteapp.R
\name{ss_deleteapp}
\alias{ss_deleteapp}
\title{Delete a deployed app}
\usage{
ss_deleteapp(session, appName, prompt = TRUE)
}
\arguments{
\item{session}{The session to connect to}

\item{appName}{The name of the app to delete}

\item{prompt}{Show remote command that will be run before deleting}
}
\description{
Delete an app that's been deployed
}
