% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_disconnect.R
\name{ss_disconnect}
\alias{ss_disconnect}
\title{Disconnect from Shiny server}
\usage{
ss_disconnect(session)
}
\arguments{
\item{session}{The session to disconnect}
}
\description{
Disconnect a session from the Shiny server
}
