% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_setupserver.R
\name{ss_does_shinyapps_exist}
\alias{ss_does_shinyapps_exist}
\title{Check whether we have a ~/ShinyApps directory on the remote account}
\usage{
ss_does_shinyapps_exist(session)
}
\arguments{
\item{session}{The session to use}
}
\description{
Check whether the ~/ShinyApps directory exists
}
