% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_frontend.R
\name{ss_frontend}
\alias{ss_frontend}
\title{Launch Shiny Sender's frontend}
\usage{
ss_frontend()
}
\description{
This launches the (optional) frontend for ShinySender
}
\details{
All of Shiny Sender's functionality can be accessed via the ss_ functions
Sometimes it's nice to do things graphically - this (locally running) Shiny
app lets you do this
}
