% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_setupserver.R
\name{ss_is_remote_package_installed}
\alias{ss_is_remote_package_installed}
\title{Check whether a package on the remote server is installed}
\usage{
ss_is_remote_package_installed(session, package)
}
\arguments{
\item{session}{The session to use}

\item{package}{The package to test for}
}
\value{
TRUE if installed, FALSE otherwise
}
\description{
Check whether a package on the remote server is installed
}
