% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_isAppNameValid.R
\name{ss_isAppNameValid}
\alias{ss_isAppNameValid}
\title{Check if a Shiny app name is valid}
\usage{
ss_isAppNameValid(appName)
}
\arguments{
\item{appName}{The name of the app}
}
\value{
TRUE if the app name is valid, FALSE otherwise
}
\description{
Valid names are alphanumeric [underscores and hyphens allowed],
between 4 and 63 characters long. Case is preserved.
https://github.com/rstudio/rsconnect/issues/142#issuecomment-261393707
}
