% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_setupserver.R
\name{ss_setupserver}
\alias{ss_setupserver}
\title{Verify server dependencies, and create `~/ShinyApps` and `~/ShinyApps_staging`
if they don't already exist}
\usage{
ss_setupserver(session)
}
\arguments{
\item{session}{ssh session, returned by `ss_connect()`}
}
\description{
Verify server dependencies, and create `~/ShinyApps` and `~/ShinyApps_staging`
if they don't already exist
}
