% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_uploadappdir.R
\name{tempBare}
\alias{tempBare}
\title{Return a string suitable for use as a temporary
filesystem}
\usage{
tempBare()
}
\description{
This is used for the staging directory of an app when we upload
one that already exists.
}
