% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_uploadappdir.R
\name{validate_url}
\alias{validate_url}
\title{Validate a URL}
\usage{
validate_url(url)
}
\arguments{
\item{url}{The URL(s) to validate}
}
\value{
TRUE if potentially valid URL, FALSE otherwise
}
\description{
Based on regex at https://cran.r-project.org/web/packages/rex/vignettes/url_parsing.html
with ftp removed
}
