% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_setupserver.R
\name{get_remote_Rprofile}
\alias{get_remote_Rprofile}
\title{Get remote .Rprofile}
\usage{
get_remote_Rprofile(session)
}
\arguments{
\item{session}{The session to use}
}
\value{
The contents of the remote .Rprofile
}
\description{
Get the remote .Rprofile on the session
}
