% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_connect.R
\name{getUserName}
\alias{getUserName}
\title{Get the login of the logged on user}
\usage{
getUserName()
}
\value{
The user name
}
\description{
Return the login of the user to logon with.  This is usually current user.
}
\details{
We can override the login by setting the SHINYSENDER_USER environment variable

The current user's login is stored in the USER and/or USERNAME environment variables
Which are populated seems to depend on the platform.  This function attempts
to get the username robustly regardless of platform

USER  USERNAME  Platform
Y    Y         Linux
Y    N         Mac
N    Y         Windows
}
