% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processls.R
\name{isShinyApp}
\alias{isShinyApp}
\title{Test if a directory contains files indicating it contains a shiny app}
\usage{
isShinyApp(directoryEntries, includeMarkdown = TRUE)
}
\arguments{
\item{directoryEntries}{A character vector containing directory entries;
one per element}

\item{includeMarkdown}{Whether to include index.Rmd as a potentially valid app}
}
\value{
TRUE if the directory contains a Shiny app, FALSE otherwise. Return
NA if an invalid combination of app files are included
}
\description{
Test for the presence of app.R or (server.R and ui.R) in a
directory listing
}
\details{
Note this doesn't test anything about the validity of the code in the
app files - it just checks if they exist
}
