% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_uploadappdir.R
\docType{data}
\name{server_proxy_overrides}
\alias{server_proxy_overrides}
\alias{prepareRprofile}
\title{Custom proxy overrides (deprecated?)}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
server_proxy_overrides

prepareRprofile(our_profile)
}
\arguments{
\item{our_profile}{Character vector containing the .Rprofile you wish to
modify}
}
\value{
A character vector containing the (potentially) modified .Rprofile
}
\description{
This function takes an .Rprofile and
modifies it with any proxy setup that is required,
using the look-up included in the package and using the
local environment variables.
}
\details{
We first test the SHINYSENDER_SERVER environment variable against the servers
specified in shinysender:::server_proxy_overrides.  If SHINYSENDER_SERVER matches
the name of any of the entries in this vector, we default to using its value as
both the http and https proxy.  This is used to handle the fact that the UoM pilot
server requires a web proxy.

This can be overriden by setting the environment variable SHINYSENDER_PROXY
(to set both http and https proxy
to the same value on the remote server),
or SHINYSENDER_HTTP_PROXY and/or SHINYSENDER_HTTPS_PROXY are set to set
 each environment variable independently.

In any case, the full URL for the proxy server(s) should be given, e.g.
"http://myproxy.com:3128"
}
\keyword{datasets}
