% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_updateRProfile.R
\name{send_Rprofile}
\alias{send_Rprofile}
\title{Send an Rprofile file to the remote server}
\usage{
send_Rprofile(session, Rprofile, appname = "~")
}
\arguments{
\item{session}{The session to use}

\item{Rprofile}{A vector containing the Rprofile to send}

\item{appname}{The location of the .Rprofile file to send. Will send .Rprofile in home directory if not specified}
}
\description{
Send an Rprofile file to the remote server
}
