% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_updateRProfile.R
\name{send_Rprofile}
\alias{send_Rprofile}
\title{Send an Rprofile file to the remote server}
\usage{
send_Rprofile(session, Rprofile)
}
\arguments{
\item{session}{The session to use}

\item{Rprofile}{A vector containing the Rprofile to send}
}
\description{
Send an Rprofile file to the remote server
}
