% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_connect.R
\name{ss_connect}
\alias{ss_connect}
\title{Connect to a Shiny Server}
\usage{
ss_connect(
  username = getUserName(),
  server = Sys.getenv("SHINYSENDER_SERVER"),
  keyfile = NULL
)
}
\arguments{
\item{username}{The user to log onto the server as. Defaults to current user}

\item{server}{The remote Shiny server to connect to.  Uses value set in
SHINYSENDER_SERVER environment variable by default}

\item{keyfile}{The path to the user's private keyfile.
the active config.  Will warn otherwise.}
}
\value{
An ssh connection object
}
\description{
Connect to the Shiny server referred to in the specified
configuration in config.yml, using the given user name
and keyfile specified
}
\note{
If keyfile is NULL, we use ssh::ssh_connects() approach to authentication
i.e. .ssh/id_rsa, followed by interactive password authentication
}
