% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_getlogs.R
\name{ss_getlogs}
\alias{ss_getlogs}
\title{Get error logs for an app from the remote server}
\usage{
ss_getlogs(session, appname = basename(getwd()), what = "last")
}
\arguments{
\item{session}{The ssh session to use}

\item{appname}{The name of the application to collect logs for.  Defaults to the name of your current local directory}

\item{what}{One of "last", "list", or a pid as returned by list.  Defaults to "last" - the most recent log for the app}
}
\description{
Get error logs for an app from the remote server
}
