% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_isAppNameValid.R
\name{ss_isAppNameValid}
\alias{ss_isAppNameValid}
\title{Check if a Shiny app name is valid}
\usage{
ss_isAppNameValid(appName)
}
\arguments{
\item{appName}{The name of the app}
}
\value{
TRUE if the app name is valid, FALSE otherwise
}
\description{
Check if a Shiny app name is valid
}
\details{
Haven't found any documentation on naming rule, so will
go strict and say alphanumeric only.

Will preserve the case the user provides.
}
