% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_setupserver.R
\name{ss_setupRprofile}
\alias{ss_setupRprofile}
\title{Set up users Rprofile so hosted apps use Packrat correctly}
\usage{
ss_setupRprofile(session)
}
\arguments{
\item{session}{The session}
}
\description{
This adds the required text to the remote ~/.Rprofile, if it
does not already exist
}
