% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_setupserver.R
\name{ss_setupserver}
\alias{ss_setupserver}
\title{Set up .Rprofile on remote server, and create ~/ShinyApps if it doesn't
exist}
\usage{
ss_setupserver(session)
}
\arguments{
\item{session}{The session to upload the .RProfile to}
}
\description{
Set up the user's environment with an appropriate .RProfile to
load packrat libraries when running their apps
}
\details{
We use Packrat to reproduce the local system's libraries as closely
as possible on the Shiny server.  We need a way of making the app
aware that it should be using its own local (i.e. Packrat) libraries
when running.

This function uploads an .Rprofile to the user's home directory
}
