% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_uploadappdir.R
\name{ss_uploadappdir}
\alias{ss_uploadappdir}
\title{Upload an app to the shiny server}
\usage{
ss_uploadappdir(session, appDir, appName)
}
\arguments{
\item{session}{The ssh session to use for the upload}

\item{appDir}{The local directory containing the app to upload}

\item{appName}{The name of the application - this will form part of the URL
for the app}
}
\description{
Upload all the files in the app directory to the Shiny server
}
